Update Dockerfile
This is a lot better than before, added back SSL capability, and user can be set to UID 999 as before. (Just needed to delgroup the PING group at 999, but that shouldn't be a problem.) Some refactoring and moving some ENV's to ARGS that can be overridden on build via comandline now. Set up for an docker-compose build with `environments:` to customize the actual build.